Объектно-ориентированное программирование (семинары). Урок 6. ООП Дизайн и Solid. Взять реализованный код в рамках семинара 4 и 5 и продемонстрировать применение принципов, усвоенных на семинаре. Нужно в проекте прокомментировать участки кода, которые рефакторим, какой принцип применяем и почему. Пример 2 (доработал)
