+++ src/main/java/OOP_Hw6/Task2/GenericMethodExample.java
@@ -4,15 +4,9 @@ package OOP_Hw6.Task2;
  * Пример использования обобщенного метода printArray для вывода элементов
  * массива любого типа.
  */
-public class GenericMethodExample {
-    /**
-     * Обобщенный метод, который принимает массив любого типа и выводит его
-     * элементы в консоль.
-     *
-     * @param array массив элементов любого типа
-     * @param <T>   обобщенный тип элементов массива
-     */
-    public static <T> void printArray(T[] array) {
+public class GenericMethodExample<T> implements ArrayPrinter<T> {
+
+    public void printArray(T[] array) {
         for (T element : array) {
             System.out.println(element);
         }
