+++ src/main/java/OOP_Hw6/Task2/OOP_Hw6_ex2.txt
@@ -2,26 +2,58 @@
 
 Комментарий к задаче 2 из ДЗ-4:
 
-1. Принцип единственной ответственности (Single Responsibility Principle, SRP): Классы и методы должны иметь только одну причину для изменения. В данном проекте каждый класс выполняет только одну конкретную задачу. Например, класс GenericMethodExample отвечает за обобщенный метод printArray, который печатает элементы массива. Класс Main содержит только метод main, который демонстрирует использование обобщенного метода. Таким образом, принцип SRP уже соблюдается.
+1. Принцип единственной ответственности (Single Responsibility Principle, SRP):
+Классы и методы должны иметь только одну причину для изменения. В данном проекте
+каждый класс выполняет только одну конкретную задачу. Например, класс
+GenericMethodExample отвечает за обобщенный метод printArray, который печатает
+элементы массива. Класс Main содержит только метод main, который демонстрирует
+использование обобщенного метода. Таким образом, принцип SRP уже соблюдается.
 
-2. Принцип открытости/закрытости (Open/Closed Principle, OCP): Классы и методы должны быть открыты для расширения, но закрыты для модификации. В данном проекте нет необходимости в изменении классов или методов для добавления новых типов массивов. Обобщенный метод printArray уже может работать с массивами любого типа, поэтому он открыт для расширения. Поэтому, принцип OCP также соблюдается.
+2. Принцип открытости/закрытости (Open/Closed Principle, OCP): Классы и методы
+должны быть открыты для расширения, но закрыты для модификации. В данном проекте
+нет необходимости в изменении классов или методов для добавления новых типов
+массивов. Обобщенный метод printArray уже может работать с массивами любого
+типа, поэтому он открыт для расширения. Поэтому, принцип OCP также соблюдается.
 
-3. Принцип подстановки Барбары Лисков (Liskov Substitution Principle, LSP): Объекты должны быть заменяемыми на свои подтипы без изменения правильности программы. В данном проекте принцип LSP не является особенно применимым, так как нет иерархии классов или подтипов, которые могли бы заменять друг друга без изменения программы.
+3. Принцип подстановки Барбары Лисков (Liskov Substitution Principle, LSP):
+Объекты должны быть заменяемыми на свои подтипы без изменения правильности
+программы. В данном проекте принцип LSP не является особенно применимым, так как
+нет иерархии классов или подтипов, которые могли бы заменять друг друга без
+изменения программы.
 
-4. Принцип разделения интерфейса (Interface Segregation Principle, ISP): Клиенты не должны зависеть от интерфейсов, которые они не используют. В данном проекте принцип ISP не применим, так как нет явного использования интерфейсов.
+4. Принцип разделения интерфейса (Interface Segregation Principle, ISP): Клиенты
+не должны зависеть от интерфейсов, которые они не используют. В данном проекте
+принцип ISP не применим, так как нет явного использования интерфейсов.
 
-5. Принцип инверсии зависимостей (Dependency Inversion Principle, DIP): Классы должны зависеть от абстракций, а не от конкретных реализаций. В данном проекте принцип DIP не применим, так как нет зависимости от конкретных реализаций классов или интерфейсов.
+5. Принцип инверсии зависимостей (Dependency Inversion Principle, DIP): Классы
+должны зависеть от абстракций, а не от конкретных реализаций. В данном проекте
+принцип DIP не применим, так как нет зависимости от конкретных реализаций
+классов или интерфейсов.
 
 Вывод: Данный код соблюдает некоторые принципы SOLID по умолчанию.
 
 Для демонстрации применения принципов SOLID внесу в код ряд изменений:
 
-1. Принцип единственной ответственности (SRP): В классе GenericMethodExample есть только один метод printArray, который выполняет только одну задачу - вывод элементов массива в консоль. Добавляю отдельный класс для чтения массива из файла, чтобы разделить ответственность за чтение и печать массива.
-2. Принцип открытости/закрытости (OCP): Для поддержки новых типов массивов, можно использовать интерфейс ArrayPrinter, который определяет метод printArray. Затем можно создать классы, реализующие этот интерфейс, для печати различных типов массивов.
+1. Принцип единственной ответственности (SRP): В классе GenericMethodExample
+есть только один метод printArray, который выполняет только одну задачу - вывод
+элементов массива в консоль. Добавляю отдельный класс для чтения массива из
+файла, чтобы разделить ответственность за чтение и печать массива.
+2. Принцип открытости/закрытости (OCP): Для поддержки новых типов массивов,
+ можно использовать интерфейс ArrayPrinter, который определяет метод printArray.
+ Затем можно создать классы, реализующие этот интерфейс, для печати различных
+ типов массивов.
 3. -
 4. -
-5. Принцип инверсии зависимостей (DIP): Можно применить принцип DIP, изменив метод main в классе Main для использования интерфейса ArrayPrinter вместо прямого вызова метода printArray класса GenericMethodExample.
+5. Принцип инверсии зависимостей (DIP): Можно применить принцип DIP, изменив
+метод main в классе Main для использования интерфейса ArrayPrinter вместо
+прямого вызова метода printArray класса GenericMethodExample.
  
-Создал отдельный интерфейс `ArrayPrinter`, который определяет метод `printArray`, и две реализации этого интерфейса - `GenericMethodExample` и `IntegerArrayPrinter`. Метод `main` теперь использует интерфейс `ArrayPrinter` для печати массивов различных типов.
+Создал отдельный интерфейс `ArrayPrinter`, который определяет метод
+`printArray`, и две реализации этого интерфейса - `GenericMethodExample` и
+`IntegerArrayPrinter`. Метод `main` теперь использует интерфейс `ArrayPrinter`
+для печати массивов различных типов.
 
-Таким образом, я продемонстрировал применение принципов SOLID в проекте. Классы и методы стали более гибкими, легко расширяемыми и поддерживаемыми, а зависимости стали более абстрактными, позволяя легко заменять конкретные реализации.
+Таким образом, я продемонстрировал применение принципов SOLID в проекте.
+Классы и методы стали более гибкими, легко расширяемыми и поддерживаемыми,
+а зависимости стали более абстрактными, позволяя легко заменять конкретные
+реализации.
